taint/test/mochitest/test_dom_parser.html

<!DOCTYPE HTML>
<html>
  <head>
    <meta charset="utf-8">
    <title>Taint DOM Parser propagation</title>
    <script src="/tests/SimpleTest/SimpleTest.js"></script>
    <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
    <script>
        function check_tainted(str) {
            ok(str.taint.length > 0, "Check tainted");
        }

        function check_untainted(str) {
            ok(str.taint.length == 0, "Check untainted");
        }

        add_task(async function test_htmlParser() {
            const parser = new DOMParser();

            const tainted_str = String.tainted("tainted=hello");
            const tainted_dom_str = "<p>" + tainted_str + "</p>";

            const doc1 = parser.parseFromString(tainted_dom_str, "text/html");

            check_tainted(doc1.body.firstChild.outerHTML);
            check_tainted(doc1.body.firstChild.textContent);

            check_tainted(doc1.body.innerHTML);
            check_tainted(doc1.body.outerHTML);
            check_tainted(doc1.body.textContent);
        });

        add_task(async function test_xmlParser() {
            const parser = new DOMParser();

            const tainted_str = String.tainted("tainted=hello");

            const tainted_xml_str = "<warning>" + tainted_str + "</warning>";
            const doc2 = parser.parseFromString(tainted_xml_str, "application/xml");

            check_tainted(doc2.documentElement.textContent);
        });

    </script>
  </head>
  <body>
    <h1>Hello world</p>
  </body>
</html>
